requirements_PI.txt
mfrc522==0.0.7
paho-mqtt==2.1.0
RPi.GPIO==0.7.1
spidev==3.8
RPLCD==1.4.0
